pyproject.toml
package-mode = false

[tool.poetry]
name = "project-template"
version = "0.1.0"
description = "Project devkit template"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-frontmatter = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
ruff = "^0.11.2"
sqlfluff = "^3.3.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
